scraping/pyproject.toml
[tool.poetry]
name = "scraping"
version = "0.1.0"
description = ""
authors = ["shraddhabhandarkar3 <bhandarkar.s@northeastern.edu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.10.15"
boto3 = "^1.28.0"
fitz = "^0.0.1.dev2" # PyMuPDF wrapper
PyMuPDF = "^1.22.5"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
pinecone-client = "^5.0.1"
openai = "^1.54.0"
python-dotenv = "^1.0.0"
tiktoken = "^0.7.0"
unstructured = {extras = ["all-docs"], version = "^0.16.8"}
langchain = "^0.3.9"
langchain-community = "^0.3.8"
langchain-openai = "^0.2.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
